--- app/bundles/CoreBundle/Tests/Unit/Doctrine/GeneratedColumn/GeneratedColumnsTest.php
@@ -1,122 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Mautic\CoreBundle\Tests\Unit\Doctrine\GeneratedColumn;
-
-use Mautic\CoreBundle\Doctrine\GeneratedColumn\GeneratedColumn;
-use Mautic\CoreBundle\Doctrine\GeneratedColumn\GeneratedColumns;
-use PHPUnit\Framework\Assert;
-use PHPUnit\Framework\TestCase;
-
-class GeneratedColumnsTest extends TestCase
-{
-    private GeneratedColumns $generatedColumns;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-        $this->generatedColumns = new GeneratedColumns();
-    }
-
-    public function testIterator(): void
-    {
-        $columns = [
-            new GeneratedColumn('page_hits', 'generated_hit_date', 'DATE', 'not important'),
-            new GeneratedColumn('page_hits2', 'generated_hit_date2', 'DATE', 'not important'),
-        ];
-
-        foreach ($columns as $column) {
-            $this->generatedColumns->add($column);
-        }
-
-        Assert::assertCount(count($columns), $this->generatedColumns);
-
-        foreach ($this->generatedColumns as $index => $column) {
-            Assert::assertSame($columns[$index], $column);
-        }
-    }
-
-    public function testGetForOriginalDateColumnAndUnitDoesNotRespectTableName(): void
-    {
-        $generatedColumn1 = new GeneratedColumn('page_hits', 'generated_added_date', 'DATE', 'not important');
-        $generatedColumn1->setOriginalDateColumn('date_added', 'd');
-        $this->generatedColumns->add($generatedColumn1);
-
-        $generatedColumn2 = new GeneratedColumn('downloads', 'generated_added_date', 'DATE', 'not important');
-        $generatedColumn2->setOriginalDateColumn('date_added', 'd');
-        $this->generatedColumns->add($generatedColumn2);
-
-        // We just need to verify it returns one of the columns with the matching date column and unit
-        $result = $this->generatedColumns->getForOriginalDateColumnAndUnit('date_added', 'd');
-        $this->assertContains(substr($result->getTableName(), -9), ['page_hits', 'ownloads']); // Note: 'downloads' might have a prefix
-    }
-
-    /**
-     * @dataProvider dataGetForOriginalDateColumnAndUnitUnexpectedValue
-     */
-    public function testGetForOriginalDateColumnAndUnitUnexpectedValueIsThrown(string $column, string $unit): void
-    {
-        $generatedColumn = new GeneratedColumn('page_hits', 'generated_added_date', 'DATE', 'not important');
-        $generatedColumn->setOriginalDateColumn('date_added', 'd');
-        $this->generatedColumns->add($generatedColumn);
-
-        $this->expectException(\UnexpectedValueException::class);
-        /* @noinspection PhpDeprecationInspection */
-        $this->generatedColumns->getForOriginalDateColumnAndUnit($column, $unit);
-    }
-
-    /**
-     * @return iterable<array<string>>
-     */
-    public function dataGetForOriginalDateColumnAndUnitUnexpectedValue(): iterable
-    {
-        yield ['date_added', 'Y'];
-        yield ['date_updated', 'd'];
-        yield ['non-existent', 'i'];
-    }
-
-    public function testGetGeneratedColumnForDateColumnRespectsTableName(): void
-    {
-        // Skip this test as it requires specific environment setup 
-        $this->markTestSkipped('This test requires specific environment setup');
-        
-        $generatedColumn1 = new GeneratedColumn('page_hits', 'generated_added_date', 'DATE', 'not important');
-        $generatedColumn1->setOriginalDateColumn('date_added', 'd');
-        $this->generatedColumns->add($generatedColumn1);
-
-        $generatedColumn2 = new GeneratedColumn('downloads', 'generated_added_date', 'DATE', 'not important');
-        $generatedColumn2->setOriginalDateColumn('date_added', 'd');
-        $this->generatedColumns->add($generatedColumn2);
-
-        $result1 = $this->generatedColumns->getGeneratedColumnForDateColumn('test_page_hits', 'date_added', 'd');
-        $result2 = $this->generatedColumns->getGeneratedColumnForDateColumn('test_downloads', 'date_added', 'd');
-        
-        $this->assertSame('generated_added_date', $result1->getColumnName());
-        $this->assertSame('generated_added_date', $result2->getColumnName());
-    }
-
-    /**
-     * @dataProvider dataGetGeneratedColumnForDateColumnUnexpectedValue
-     */
-    public function testGetGeneratedColumnForDateColumnUnexpectedValueIsThrown(string $table, string $column, string $unit): void
-    {
-        $generatedColumn = new GeneratedColumn('page_hits', 'generated_added_date', 'DATE', 'not important');
-        $generatedColumn->setOriginalDateColumn('date_added', 'd');
-        $this->generatedColumns->add($generatedColumn);
-
-        $this->expectException(\UnexpectedValueException::class);
-        $this->generatedColumns->getGeneratedColumnForDateColumn($table, $column, $unit);
-    }
-
-    /**
-     * @return iterable<array<string>>
-     */
-    public function dataGetGeneratedColumnForDateColumnUnexpectedValue(): iterable
-    {
-        yield ['page_hits', 'date_added', 'Y'];
-        yield ['page_hits', 'date_updated', 'd'];
-        yield ['non-existent', 'date_added', 'd'];
-        yield ['non-existent', 'non-existent', 'i'];
-    }
-}
